+++ cmd/serve.go
@@ -22,7 +22,7 @@ var serveCmd = &cobra.Command{
 	Use:   "serve",
 	Short: "Start a REST API server",
 	Long: `Start a REST API server for accessing and manipulating the catalog.
-A target repository for push operations must be specified with --pushTarget in case neither --repo nor --directory is given.
+A target repository for import operations must be specified with --importTarget in case neither --repo nor --directory is given.
 This may be omitted if there's exactly one repository configured`,
 	Args: cobra.MaximumNArgs(0),
 	Run:  serve,
@@ -36,8 +36,8 @@ func init() {
 	_ = serveCmd.RegisterFlagCompletionFunc("repo", completion.CompleteRepoNames)
 	serveCmd.Flags().StringP("directory", "d", "", "Use the specified directory as repository. This option allows directly using a directory as a local TM repository, forgoing creating a named repository.")
 	_ = serveCmd.MarkFlagDirname("directory")
-	serveCmd.Flags().StringP("pushTarget", "t", "", "Name of the repo to use as target for push operations")
-	_ = serveCmd.RegisterFlagCompletionFunc("pushTarget", completion.CompleteRepoNames)
+	serveCmd.Flags().StringP("importTarget", "t", "", "Name of the repo to use as target for import operations")
+	_ = serveCmd.RegisterFlagCompletionFunc("importTarget", completion.CompleteRepoNames)
 	serveCmd.Flags().String(config.KeyUrlContextRoot, "",
 		"Define additional URL context root path to be considered in hypermedia links (env var TMC_URLCONTEXTROOT)")
 	serveCmd.Flags().String(config.KeyCorsAllowedOrigins, "", "Set comma-separated list for CORS allowed origins (env var TMC_CORSALLOWEDORIGINS)")
@@ -63,7 +63,7 @@ func serve(cmd *cobra.Command, args []string) {
 	port := cmd.Flag("port").Value.String()
 	repo := cmd.Flag("repo").Value.String()
 	dir := cmd.Flag("directory").Value.String()
-	pushTarget := cmd.Flag("pushTarget").Value.String()
+	importTarget := cmd.Flag("importTarget").Value.String()
 	spec, err := model.NewSpec(repo, dir)
 	if errors.Is(err, model.ErrInvalidSpec) {
 		cli.Stderrf("Invalid specification of repository to be served. --repo and --directory are mutually exclusive. Set at most one")
@@ -72,11 +72,11 @@ func serve(cmd *cobra.Command, args []string) {
 
 	opts := getServeOptions()
 
-	pushSpec := spec
-	if repo == "" && dir == "" && pushTarget != "" {
-		pushSpec = model.NewRepoSpec(pushTarget)
+	importSpec := spec
+	if repo == "" && dir == "" && importTarget != "" {
+		importSpec = model.NewRepoSpec(importTarget)
 	}
-	err = cli.Serve(host, port, opts, spec, pushSpec)
+	err = cli.Serve(host, port, opts, spec, importSpec)
 	if err != nil {
 		cli.Stderrf("serve failed")
 		os.Exit(1)