+++ internal/repos/fs_test.go
@@ -190,9 +190,9 @@ func TestFileRepo_Push(t *testing.T) {
 	}
 	tmName := "omnicorp-tm-department/omnicorp/omnilamp"
 	id := tmName + "/v0.0.0-20231208142856-c49617d2e4fc.tm.json"
-	_, err := r.Push(context.Background(), model.MustParseTMID(id), []byte{}, PushOptions{})
+	_, err := r.Import(context.Background(), model.MustParseTMID(id), []byte{}, ImportOptions{})
 	assert.Error(t, err)
-	_, err = r.Push(context.Background(), model.MustParseTMID(id), []byte("{}"), PushOptions{})
+	_, err = r.Import(context.Background(), model.MustParseTMID(id), []byte("{}"), ImportOptions{})
 	assert.NoError(t, err)
 	assert.FileExists(t, filepath.Join(temp, id))
 
@@ -202,21 +202,21 @@ func TestFileRepo_Push(t *testing.T) {
 	_ = os.WriteFile(filepath.Join(temp, tmName, "v0.0.1-20231208142856-d49617d2e4fc.tm.json"), []byte("{}"), defaultFilePermissions)
 
 	id2 := "omnicorp-tm-department/omnicorp/omnilamp/v1.0.0-20231219123456-a49617d2e4fc.tm.json"
-	res, err := r.Push(context.Background(), model.MustParseTMID(id2), []byte("{}"), PushOptions{})
+	res, err := r.Import(context.Background(), model.MustParseTMID(id2), []byte("{}"), ImportOptions{})
 	expCErr := &ErrTMIDConflict{Type: IdConflictSameContent, ExistingId: "omnicorp-tm-department/omnicorp/omnilamp/v1.0.0-20231208142856-a49617d2e4fc.tm.json"}
 	assert.Equal(t, expCErr, err)
-	assert.Equal(t, PushResult{Type: PushResultTMExists, Message: expCErr.Error(), Err: expCErr}, res)
+	assert.Equal(t, ImportResult{Type: ImportResultTMExists, Message: expCErr.Error(), Err: expCErr}, res)
 
 	id3 := "omnicorp-tm-department/omnicorp/omnilamp/v1.0.0-20231219123456-f49617d2e4fc.tm.json"
-	_, err = r.Push(context.Background(), model.MustParseTMID(id3), []byte("{}"), PushOptions{})
+	_, err = r.Import(context.Background(), model.MustParseTMID(id3), []byte("{}"), ImportOptions{})
 	assert.NoError(t, err)
 	assert.FileExists(t, filepath.Join(temp, id3))
 
 	id4 := "omnicorp-tm-department/omnicorp/omnilamp/v1.0.0-20231219123456-049617d2e4fc.tm.json"
-	res, err = r.Push(context.Background(), model.MustParseTMID(id4), []byte("{\"val\":1}"), PushOptions{})
+	res, err = r.Import(context.Background(), model.MustParseTMID(id4), []byte("{\"val\":1}"), ImportOptions{})
 	assert.NoError(t, err)
 	expCErr = &ErrTMIDConflict{Type: IdConflictSameTimestamp, ExistingId: id3}
-	assert.Equal(t, PushResult{Type: PushResultWarning, TmID: id4, Message: expCErr.Error(), Err: expCErr}, res)
+	assert.Equal(t, ImportResult{Type: ImportResultWarning, TmID: id4, Message: expCErr.Error(), Err: expCErr}, res)
 
 }