+++ internal/repos/http.go
@@ -78,8 +78,8 @@ func newBaseHttpRepo(config map[string]any, spec model.RepoSpec) (baseHttpRepo, 
 	return base, nil
 }
 
-func (h *HttpRepo) Push(ctx context.Context, id model.TMID, raw []byte, opts PushOptions) (PushResult, error) {
-	return PushResult{}, ErrNotSupported
+func (h *HttpRepo) Import(ctx context.Context, id model.TMID, raw []byte, opts ImportOptions) (ImportResult, error) {
+	return ImportResult{}, ErrNotSupported
 }
 func (h *HttpRepo) Delete(ctx context.Context, id string) error {
 	return ErrNotSupported